+++ src/index.ts
@@ -174,9 +174,9 @@ export async function runTest() {
   console.log("Running test: " + test.name)
   console.log("─".repeat(test.name.length + 20))
 
-  let isTsFile = !runWithBun && /\.ts$/.test(test.fullPath)
+  let isTsFile = /\.ts$/.test(test.fullPath)
   let filePath = test.fullPath
-  if (isTsFile) {
+  if (!runWithBun && isTsFile) {
     if (configPath === "") {
       console.error("tsconfig.json not found at: " + configPath)
       return
@@ -203,7 +203,7 @@ export async function runTest() {
       }
     })
   })
-  if (isTsFile) {
+  if (!runWithBun && isTsFile) {
     fs.unlinkSync(filePath)
   }
   console.log(" ")
